infra/searchfuncapp/requirements.txt
azure-functions==1.19.0
azure-search-documents==11.5.3
azure-core==1.30.2
